Crate flowrgui

flowrgui is a GUI flow runner for running flow programs.

It reads a compiled FlowManifest produced by a flow compiler, such as flowc, that describes the graph of communicating functions that constitute the flow program.

Use flowrgui --help or flowrgui -h at the command line to see the command line options

The gui module implements a set of context functions, adapted to a GUI runner that allow the flow program to interact with the environment where it is being run.

Depending on the command line options supplied flowrgui executes the Coordinator of flow execution in a background thread or connects to an already running coordinator in another process. Application and Coordinator (thread or process) communicate via network messages using the SubmissionHandler to submit flows for execution, and interchanging ClientMessages and CoordinatorMessages for execution of context interaction in the client, as requested by functions running in the coordinator’s Executors

Modules

context 🔒

Include the module that implements the context functions

coordinator 🔒

module that runs a coordinator in background

errors 🔒

provides Error that other modules in this crate will use crate::errors::*; to get access to everything error_chain creates.

gui 🔒

provides the context functions for interacting with the execution environment from a flow, plus client-Coordinator implementations of flowrlib::submission_handler for executing them on different threads from the Coordinator

Structs

FlowrGui 🔒

ImageReference 🔒

ServerSettings

Settings to use when starting a coordinator server

SubmissionSettings 🔒

UiSettings 🔒

Enums

CoordinatorSettings

CoordinatorSettings captures the parameters to be used when creating a new Coordinator

CoordinatorState 🔒

Message

Message enum captures all the types of messages that are sent to and processed by the FlowrGui Iced Application

Statics

STDOUT_SCROLLABLE_ID 🔒

Functions

main 🔒

Main for flowrgui binary - call run() and print any error that results or exit silently if OK