Enum flowrgui::Message

pub enum Message {
    CoordinatorDisconnected,
    CoordinatorSent(CoordinatorMessage),
    SubmitFlow,
    Submitted,
    UrlChanged(String),
    FlowArgsChanged(String),
    TabSelected(usize),
    StdioAutoScrollTogglerChanged(bool),
}

Message enum captures all the types of messages that are sent to and processed by the FlowrGui Iced Application

Variants

CoordinatorDisconnected

We lost contact with the coordinator

CoordinatorSent(CoordinatorMessage)

The Coordinator sent to the client/App a Coordinator Message

SubmitFlow

The UI has requested to submit the flow to the Coordinator for execution

Submitted

ddd

UrlChanged(String)

The Url of the flow to run has been edited by the UI

FlowArgsChanged(String)

The arguments to send to the flow when executed have been edited by the UI

TabSelected(usize)

A different tab of stdio has been selected

StdioAutoScrollTogglerChanged(bool)

The toggle to auto-scroll to bottom of STDIO has changed