Enum flowrgui::gui::coordinator_message::CoordinatorMessage

pub enum CoordinatorMessage {

    Connected(Sender<ClientMessage>),
    FlowStart,
    FlowEnd(Metrics),
    CoordinatorExiting(Result<()>),
    Stdout(String),
    Stderr(String),
    GetStdin,
    GetLine(String),
    GetArgs,
    Read(String),
    Write(String, Vec<u8>),
    PixelWrite((u32, u32), (u8, u8, u8), (u32, u32), String),
    StdoutEof,
    StderrEof,
    Invalid,

}

An Message sent from the runtime server to a runtime_client

Variants

Connected(Sender<ClientMessage>)

** These messages are used to communicate to the app the connection status to the Coordinator A connection has been made

FlowStart

** These messages are used to implement the SubmissionProtocol between the coordinator and the cli_client A flow has started executing

FlowEnd(Metrics)

A flow has stopped executing

CoordinatorExiting(Result<()>)

Coordinator is exiting, with a result (OK, or Err)

Stdout(String)

** These messages are used to implement the context functions between the cli_runtime_server that runs as part of the Coordinator and the cli_runtime_client that interacts with STDIO A String of contents was sent to stdout

Stderr(String)

A String of contents was sent to stderr

GetStdin

A Request to read from Stdin

GetLine(String)

A Request to read a line of characters from Stdin, with a String prompt to show

GetArgs

A Request to get the arguments for the flow

Read(String)

A Request to read bytes from a file

Write(String, Vec<u8>)

A Request to write a series of bytes to a file

PixelWrite((u32, u32), (u8, u8, u8), (u32, u32), String)

A Request to write a pixel to an ImageBuffer

StdoutEof

A Request to snd EOF to Stdout

StderrEof

A Request to snd EOF to Stderr

Invalid

Invalid - used when deserialization goes wrong