Struct flowrgui::gui::debug_client::DebugClient

pub struct DebugClient {
    connection: ClientConnection,
    override_args: Arc<Mutex<Vec<String>>>,
    editor: Editor<(), DefaultHistory>,
    last_command: String,
}

Fields
connection: ClientConnectionoverride_args: Arc<Mutex<Vec<String>>>editor: Editor<(), DefaultHistory>last_command: String
Implementations

impl DebugClient

pub fn new( connection: ClientConnection, override_args: Arc<Mutex<Vec<String>>> ) -> Self

Create a new debug client accepting the debug connection

pub fn debug_client_loop(self)

Main debug client loop where events are received, processed and responses sent

fn help()

fn parse_command( &self, input: String ) -> Result<(String, String, Option<Vec<String>>)>

fn parse_optional_int(params: Option<Vec<String>>) -> Option<usize>

fn parse_breakpoint_spec(specs: Option<Vec<String>>) -> Option<BreakpointSpec>

fn parse_inspect_spec(spec: Option<Vec<String>>) -> Option<DebugCommand>

fn get_user_command(&mut self, job_number: usize) -> Result<DebugCommand>

fn get_server_command( &mut self, command: &str, params: Option<Vec<String>> ) -> Option<DebugCommand>

fn process_server_message( &mut self, message: DebugServerMessage ) -> Result<DebugCommand>

fn function_list(functions: Vec<RuntimeFunction>)

fn display_state(run_state: &RunState)